REFACTOR add and remove prices
Before add and remove were by each price, It meaning each add and remove
operation made a request to database and It did slow the operations. Now
all operations is in frontend and when all operation(add or remove) are
finished then recent the user can save(make a request to db). It makes
very speed behavior face to user.